[SPARK-30336][SQL][SS] Move Kafka consumer-related classes to its own package
